requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
XlsxWriter==3.0.3
psycopg2==2.9.1
python-dotenv==0.15.0
djangorestframework==3.14.0
